refactor: currencyDecorator move to util
순수함수는 util로 빼도록 결정함으로써 수정
